requirements.txt
functions-framework
google-cloud-discoveryengine
google-cloud-asset
google-api-python-client

